added dispatchers removing GlobalScope
