+++ app/src/main/java/luci/sixsixsix/powerampache2/PowerAmpache2Application.kt
@@ -23,42 +23,25 @@ package luci.sixsixsix.powerampache2
 
 import android.app.Application
 import android.content.Context
-import android.util.Log
 import androidx.work.Configuration
-import androidx.work.ListenableWorker
-import androidx.work.WorkerFactory
-import androidx.work.WorkerParameters
 import coil.ImageLoader
 import coil.ImageLoaderFactory
-import coil.disk.DiskCache
-import coil.memory.MemoryCache
-import coil.request.CachePolicy
 import dagger.hilt.android.HiltAndroidApp
 import luci.sixsixsix.powerampache2.domain.common.Constants
-
-import luci.sixsixsix.powerampache2.data.local.MusicDatabase
-import luci.sixsixsix.powerampache2.data.remote.MainNetwork
-import luci.sixsixsix.powerampache2.data.remote.worker.SongDownloadWorker
-import luci.sixsixsix.powerampache2.di.AmpacheOkHttpClientBuilder
-import luci.sixsixsix.powerampache2.domain.common.Constants.TIMEOUT_CONNECTION_S
-import luci.sixsixsix.powerampache2.domain.common.Constants.TIMEOUT_READ_S
-import luci.sixsixsix.powerampache2.domain.common.Constants.TIMEOUT_WRITE_S
 import luci.sixsixsix.powerampache2.domain.utils.ConfigProvider
-import luci.sixsixsix.powerampache2.domain.utils.StorageManager
 import org.acra.config.mailSender
 import org.acra.data.StringFormat
 import org.acra.ktx.initAcra
-import java.util.concurrent.TimeUnit
 import javax.inject.Inject
 
 @HiltAndroidApp
 class PowerAmpache2Application : Application(), ImageLoaderFactory, Configuration.Provider {
 
     @Inject
-    lateinit var workerFactory: SongDownloadWorkerFactory
+    lateinit var workerFactoryConfiguration: Configuration
 
     @Inject
-    lateinit var imageLoaderOkHttpClient: AmpacheOkHttpClientBuilder
+    lateinit var imageLoaderBuilder: ImageLoader.Builder
 
     @Inject
     lateinit var configProvider: ConfigProvider
@@ -92,63 +75,11 @@ class PowerAmpache2Application : Application(), ImageLoaderFactory, Configuratio
         }
     }
 
-    /**
-     * TODO: move to appmodule and inject with hilt
-     */
-    override fun newImageLoader(): ImageLoader = ImageLoader(this).newBuilder()
-        .crossfade(200)
+    override fun newImageLoader(): ImageLoader = imageLoaderBuilder
         .placeholder(R.drawable.placeholder_album)
         .error(R.drawable.placeholder_album)
-        .diskCachePolicy(CachePolicy.ENABLED)
-        .memoryCachePolicy(CachePolicy.ENABLED)
-        .networkCachePolicy(CachePolicy.ENABLED)
-        .okHttpClient(
-            imageLoaderOkHttpClient(true)
-                //.retryOnConnectionFailure(true)
-                .connectTimeout(TIMEOUT_CONNECTION_S, TimeUnit.SECONDS)
-                .readTimeout(TIMEOUT_READ_S, TimeUnit.SECONDS)
-                .writeTimeout(TIMEOUT_WRITE_S, TimeUnit.SECONDS)
-                .build()
-        )
-        //.respectCacheHeaders(false)
-        .memoryCache {
-            MemoryCache.Builder(this)
-                .maxSizePercent(0.08)
-                .strongReferencesEnabled(false)
-                .build()
-        }
-        .diskCache {
-            DiskCache.Builder()
-                .maxSizePercent(0.10)
-                .directory(getDir("paimages", MODE_PRIVATE))
-                .build()
-
-        }
-        //.logger(DebugLogger()) // TODO change in production
         .build()
 
     override val workManagerConfiguration: Configuration
-        get() = Configuration.Builder()
-            .setMinimumLoggingLevel(Log.INFO)
-            .setWorkerFactory(workerFactory)
-            .build()
-
-}
-
-class SongDownloadWorkerFactory @Inject constructor(
-    private val api: MainNetwork,
-    private val db: MusicDatabase,
-    private val storageManager: StorageManager
-): WorkerFactory() {
-    override fun createWorker(
-        appContext: Context,
-        workerClassName: String,
-        workerParameters: WorkerParameters
-    ): ListenableWorker = SongDownloadWorker(
-        api,
-        db,
-        storageManager,
-        appContext,
-        workerParameters
-    )
+        get() = workerFactoryConfiguration
 }
