+++ app/src/main/java/luci/sixsixsix/powerampache2/data/BaseAmpacheRepository.kt
@@ -24,6 +24,7 @@ package luci.sixsixsix.powerampache2.data
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.distinctUntilChanged
 import androidx.lifecycle.map
+import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.FlowCollector
 import kotlinx.coroutines.flow.catch
@@ -277,6 +278,10 @@ abstract class BaseAmpacheRepository(
             return false
         }
         return true
+    }
+
+    suspend fun clearSession() {
+        dao.clearSession()
     }
 }
