+++ app/src/main/java/luci/sixsixsix/powerampache2/presentation/screens/main/AuthViewModel.kt
@@ -43,7 +43,7 @@ import luci.sixsixsix.mrlog.L
 import luci.sixsixsix.powerampache2.R
 import luci.sixsixsix.powerampache2.common.Resource
 import luci.sixsixsix.powerampache2.domain.common.sha256
-import luci.sixsixsix.powerampache2.data.remote.disableSSLCertificateVerify
+import luci.sixsixsix.powerampache2.common.disableSSLCertificateVerify
 import luci.sixsixsix.powerampache2.domain.MusicRepository
 import luci.sixsixsix.powerampache2.domain.utils.AlarmScheduler
 import luci.sixsixsix.powerampache2.domain.utils.SharedPreferencesManager
