+++ gradle/libs.versions.toml
@@ -16,7 +16,7 @@ material3 = "1.3.2"
 materialIconsExtended = "1.7.8"
 media = "1.7.0"
 media3 = "1.6.1"
-okhttpVersion = "5.0.0-alpha.14"
+okhttpVersion = "4.12.0" #"5.0.0-alpha.14"
 pagingCompose = "3.3.6"
 retrofit = "2.11.0"
 roomVersion = "2.7.1"
